turn the scope resolution around, to the proper way.
I used to check whether there is a field in the scope with the same name
as the variable. If yes, I would prefix "$scope.", if not I wouldn't.

Now by default I put "$scope.". But I keep track of variables which get
generated, and if I find the variable name in the generated ones, then
I won't put "$scope.".

Advantages:
* if a local variable name would shadow a field from the scope, we would
  still wrongly prefix with "$scope."
* scope inheritance is now much more manageable because we don't have to
  track all the parent interfaces and their fields to resolve the $scope
* we will have compile errors now if a variable is neither in the scope
  nor declared in local variables. That used to build.
* we now mimic exactly what angular is doing, which is much more healthy
  and intuitive.